requirements.txt
Django>=4.0
djangorestframework
djangorestframework-simplejwt
python-dotenv
celery
redis
python-telegram-bot==20.8